+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Add support for specifying additional extensions to load ([#547], [#563])
+
+[#547]: https://github.com/stackabletech/druid-operator/pull/547
+[#563]: https://github.com/stackabletech/druid-operator/pull/563
+
 ## [24.3.0] - 2024-03-20
 
 ### Added